refactor: use common category component
