+++ next/web/src/App/Admin/Settings/Automations/actions/UpdateCategoryId.tsx
@@ -2,7 +2,7 @@ import { Controller, useFormContext } from 'react-hook-form';
 import { get } from 'lodash-es';
 
 import { Form } from '@/components/antd';
-import { CategorySelect } from '../components/CategorySelect';
+import { CategorySelect } from '@/components/common';
 
 export function UpdateCategoryId({ path }: { path: string }) {
   const { control, formState } = useFormContext();
@@ -14,7 +14,7 @@ export function UpdateCategoryId({ path }: { path: string }) {
         control={control}
         name={`${path}.value`}
         rules={{ required: true }}
-        render={({ field }) => <CategorySelect {...field} initValue={field.value} />}
+        render={({ field }) => <CategorySelect {...field} style={{ width: 260 }} />}
       />
     </Form.Item>
   );
